+++ libmemdraw/unload.c
@@ -17,7 +17,7 @@ unloadmemimage(Memimage *i, Rectangle r, uchar *data, int ndata)
 	q = byteaddr(i, r.min);
 	for(y=r.min.y; y<r.max.y; y++){
 		memmove(data, q, l);
-		q += i->width*sizeof(ulong);
+		q += i->width*sizeof(u32);
 		data += l;
 	}
 	return ndata;
